feat: card component 모바일 스타일 반영 완료
