+++ pages/_app.tsx
@@ -105,9 +105,7 @@ function MyApp({ Component, pageProps }: AppProps) {
           },
           '@mobile': {
             marginTop: '70px',
-          },
-          '@small_mobile': {
-            px: '$12',
+            px: '$16',
           },
         }}
       >
