+++ src/components/util/Divider.tsx
@@ -1,0 +1,9 @@
+import { Box } from '@components/box/Box';
+import { styled } from '@stitches/react';
+
+export const Divider = styled(Box, {
+  width: '100%',
+  borderRadius: '1px',
+  height: '1px',
+  backgroundColor: '$black80',
+});
